reporting/bufferSize.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>Reporting: Buffer size</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
window.onload = function() {
  async_test(function(test) {
    var observer = new ReportingObserver(function(reports, observer) {
      test.step(function() {
        assert_equals(reports.length, 100);
        for(i = 0; i != 100; ++i)
        assert_equals(reports[i].body.message, "" + (i + 50));
      });

      test.done();
    });

    for (i = 0; i != 150; ++i) {
      let report = new CustomReport("" + i);
      report.send();
    }

    observer.observe();
  });
};
</script>
</body>
</html>
